code cleanups and minor fixes
